python-backend/requirements.txt
llama-index==0.12.47
llama-index-llms-groq==0.3.2
jedi==0.19.2
llama-index-embeddings-huggingface==0.5.5
llama-index-vector-stores-chroma==0.4.2
gradio==5.35.0
python-dotenv==1.1.1
chromadb==1.0.15
pymysql==1.1.1
rapidfuzz==3.13.0
llama-index-llms-google-genai==0.2.3
sse-starlette==2.4.1
